# Chore: Replace some function application with spread operators