--- a/lib/rules/no-useless-return.js
+++ b/lib/rules/no-useless-return.js
@@ -16,15 +16,6 @@ const astUtils = require("../ast-utils"),
 //------------------------------------------------------------------------------
 
 /**
- * Adds all elements of 2nd argument into 1st argument.
- *
- * @param {Array} array - The destination array to add.
- * @param {Array} elements - The source array to add.
- * @returns {void}
- */
-const pushAll = Function.apply.bind(Array.prototype.push);
-
-/**
  * Removes the given element from the array.
  *
  * @param {Array} array - The source array to remove.
@@ -137,7 +128,7 @@ module.exports = {
                     continue;
                 }
 
-                pushAll(uselessReturns, segmentInfoMap.get(segment).uselessReturns);
+                uselessReturns.push(...segmentInfoMap.get(segment).uselessReturns);
             }
 
             return uselessReturns;