@platinumazure 2018-07-22
I ran some performance tests just to see if anything showed up. Seems negligible.

Before:

```

> eslint@5.2.0 perf C:\Users\Kevin\Documents\GitHub\eslint
> node Makefile.js perf


Loading:
  Load performance Run #1:  481.270948ms
  Load performance Run #2:  485.966138ms
  Load performance Run #3:  494.304693ms
  Load performance Run #4:  481.03489ms
  Load performance Run #5:  486.99524ms

  Load Performance median:  485.966138ms


Single File:
  CPU Speed is 2208 with multiplier 13000000
  Performance Run #1:  12369.628925ms
  Performance Run #2:  12088.161432ms
  Performance Run #3:  12014.560007ms
  Performance Run #4:  12105.00828ms
  Performance Run #5:  11961.682069ms

  Performance budget exceeded: 12088.161432ms (limit: 5887.68115942029ms)


Multi Files (0 files):
  CPU Speed is 2208 with multiplier 39000000
  Performance Run #1:  29699.744325ms
  Performance Run #2:  29950.80021ms
  Performance Run #3:  29290.550859ms
  Performance Run #4:  29211.821083ms
  Performance Run #5:  30003.076175ms

  Performance budget exceeded: 29699.744325ms (limit: 17663.043478260868ms)
```

After:

```
> eslint@5.2.0 perf C:\Users\Kevin\Documents\GitHub\eslint
> node Makefile.js perf


Loading:
  Load performance Run #1:  484.968109ms
  Load performance Run #2:  481.33773ms
  Load performance Run #3:  490.3993ms
  Load performance Run #4:  480.28544ms
  Load performance Run #5:  486.001849ms

  Load Performance median:  484.968109ms


Single File:
  CPU Speed is 2208 with multiplier 13000000
  Performance Run #1:  12330.415922ms
  Performance Run #2:  11991.310372ms
  Performance Run #3:  12168.240774ms
  Performance Run #4:  11979.039527ms
  Performance Run #5:  12050.532662ms

  Performance budget exceeded: 12050.532662ms (limit: 5887.68115942029ms)


Multi Files (0 files):
  CPU Speed is 2208 with multiplier 39000000
  Performance Run #1:  29812.813363ms
  Performance Run #2:  29327.754355ms
  Performance Run #3:  29198.634759ms
  Performance Run #4:  29909.118566ms
  Performance Run #5:  30264.074325ms

  Performance budget exceeded: 29812.813363ms (limit: 17663.043478260868ms)
```